Jit: merge duplicate hash code
